+++ src/docstrings.jl
@@ -33,18 +33,17 @@ Lattice{Float64,3,2} : 2D lattice in 3D space
 ```
 
 # See also
-    `RegionPresets`, `HamiltonianPresets`
+    `RegionPresets`, `HamiltonianPresets`, `ExternalPresets`
 """
 LatticePresets
 
 """
 `HamiltonianPresets` is a Quantica submodule containing several pre-defined Hamiltonians.
-The alias `HP` can be used in place of `LatticePresets`. Currently supported hamiltonians
-are
+The alias `HP` can be used in place of `HamiltonianPresets`. Currently supported
+hamiltonians are
 
     HP.graphene(; kw...)
     HP.twisted_bilayer_graphene(; kw...)
-    wannier90(file, model...; kw...)
 
 For details on the keyword arguments `kw` see the corresponding docstring
 
@@ -61,7 +60,7 @@ Hamiltonian{Float64,3,2}: Hamiltonian on a 2D Lattice in 3D space
 ```
 
 # See also
-    `LatticePresets`, `RegionPresets`
+    `LatticePresets`, `RegionPresets`, `ExternalPresets`
 
 """
 HamiltonianPresets
@@ -99,9 +98,25 @@ true
 ```
 
 # See also
-    `LatticePresets`, `HamiltonianPresets`
+    `LatticePresets`, `HamiltonianPresets`, `ExternalPresets`
 """
 RegionPresets
+
+"""
+`ExternalPresets` is a Quantica submodule containing utilities to import objects from
+external applications The alias `EP` can be used in place of `ExternalPresets`. Currently
+supported importers are
+
+    EP.wannier90(args...; kw...)
+
+For details on the arguments `args...` and keyword arguments `kw` see the corresponding
+docstring.
+
+# See also
+    `LatticePresets`, `RegionPresets`, `HamiltonianPresets`
+
+"""
+ExternalPresets
 
 """
     sublat(sites...; name::Symbol = :A)
@@ -216,14 +231,6 @@ lattice `lat[sel...]` with some `siteselector` keywords `sel`. See also `lattice
 
 Note: the returned collections can be of different types (vectors, generators, views...)
 
-    sites(b::WannierBuilder)
-
-Returns the Wannier position matrix in the form of a `r::BarebonesOperator` object, which
-can be indexed as `r[s, s´]` to obtain matrix elements `⟨s|R|s´⟩` of the position operator
-`R` (a vector). Here `s` and `s´` represent site indices, constructued with `cellsites`. To
-obtain the matrix between cells separated by `dn::SVector{L,Int}`, do `r[dn]`. The latter
-will throw an error if the `dn` harmonic is not present.
-
 # Examples
 ```jldoctest
 julia> sites(LatticePresets.honeycomb(), :A)
@@ -235,6 +242,17 @@ julia> sites(LatticePresets.honeycomb(), :A)
     `lattice`, `siteselector`, `cellsites`
 """
 sites
+
+"""
+    position(b::ExternalPresets.WannierBuilder)
+
+Returns the Wannier position matrix in the form of a `r::BarebonesOperator` object, which
+can be indexed as `r[s, s´]` to obtain matrix elements `⟨s|R|s´⟩` of the position operator
+`R` (a vector). Here `s` and `s´` represent site indices, constructued with `cellsites`. To
+obtain the matrix between cells separated by `dn::SVector{L,Int}`, do `r[dn]`. The latter
+will throw an error if the `dn` harmonic is not present.
+"""
+position
 
 """
     supercell(lat::Lattice{E,L}, v::NTuple{L,Integer}...; seed = missing, kw...)
@@ -2064,62 +2082,3 @@ is done more efficiently internally.
 
 """
 diagonal
-
-"""
-    wannier90(filename::String; kw...)
-
-Import a Wannier90 tight-binding file in the form of a `w::WannierBuilder` object. It can be
-used to obtain a `Hamiltonian` with `hamiltonian(w)`, and the matrix of the position
-operator with `sites(w)`.
-
-    wannier90(filename, model::AbstractModel; kw...)
-
-Modify the `WannierBuilder` after import by adding `model` to it.
-
-    push!(w::WannierBuilder, modifier::AbstractModifier)
-    w |> modifier
-
-Applies a `modifier` to `w`.
-
-## Keywords
-- htol : skip matrix elements of the Hamiltonian smaller than this (in absolute value). Default: `1e-8`
-- rtol : skip non-diagonal matrix elements of the position operator smaller than this (in absolute value). Default: `1e-8`
-- dim : override the dimensionality of the imported system, dropping trailing dimensions beyond `dim`. Default: `3`
-- type: override the real number type of the imported system. Default: `Float64`
-
-# Examples
-```
-julia> w = wannier90("wannier_tb.dat", @onsite((; o) -> o);  htol = 1e-4, rtol = 1e-4, dim = 2, type = Float32)
-WannierBuilder{Float32,2} : 2-dimensional Hamiltonian builder of type Float32 from Wannier90 input
-  cells      : 151
-  elements   : 6724
-  modifiers  : 1
-
-julia> h = hamiltonian(w)
-ParametricHamiltonian{Float32,2,2}: Parametric Hamiltonian on a 2D Lattice in 2D space
-  Bloch harmonics  : 151
-  Harmonic size    : 10 × 10
-  Orbitals         : [1]
-  Element type     : scalar (ComplexF32)
-  Onsites          : 10
-  Hoppings         : 6704
-  Coordination     : 670.4
-  Parameters       : [:o]
-
-julia> r = sites(w)
-BarebonesOperator{2}: a simple collection of 2D Bloch harmonics
-  Bloch harmonics  : 151
-  Harmonic size    : 10 × 10
-  Element type     : SVector{2, ComplexF32}
-  Nonzero elements : 7408
-
-julia> r[cellsites(SA[0,0], 3), cellsites(SA[1,0],2)]
-2-element SVector{2, ComplexF32} with indices SOneTo(2):
- -0.0016230071f0 - 0.00012927242f0im
-   0.008038711f0 + 0.004102786f0im
-```
-
-# See also
-    `hamiltonian`, `sites`
-"""
-wannier90
